Break bytesToHumanSize to lib/utils
